routingCalculation: Created RoutingCalculator module in transition-backend to encapsulate routing and accessibility map endpoint logic
