+++ packages/transition-backend/src/api/__tests__/public.routes.test.ts
@@ -11,16 +11,10 @@ import request from 'supertest';
 import transitObjectDataHandlers from '../../services/transitObjects/TransitObjectsDataHandler';
 import osrmProcessManager from 'chaire-lib-backend/lib/utils/processManagers/OSRMProcessManager';
 import tokensDbQueries from 'chaire-lib-backend/lib/models/db/tokens.db.queries';
-import trRoutingProcessManager from 'chaire-lib-backend/lib/utils/processManagers/TrRoutingProcessManager';
-import { TransitAccessibilityMapCalculator } from 'transition-common/lib/services/accessibilityMap/TransitAccessibilityMapCalculator';
-import { TransitRoutingCalculator } from 'transition-common/lib/services/transitRouting/TransitRoutingCalculator';
 import TrError from 'chaire-lib-common/lib/utils/TrError';
+import { calculateAccessibilityMap, calculateRoute } from '../../services/routingCalculation/RoutingCalculator';
 
-// Required to test POST /api/accessibility endpoint
-jest.mock('transition-common/lib/services/nodes/NodeCollection');
-
-// Required to test POST /api/route endpoint
-jest.mock('transition-common/lib/services/path/PathCollection');
+jest.mock('../../services/routingCalculation/RoutingCalculator');
 
 // Mock passport (therefore ignoring authentication)
 jest.mock('passport');
@@ -123,229 +117,84 @@ describe('Testing endpoints', () => {
     });
 
     test('POST /api/accessibility, without geojson', async () => {
-        const routingResult = 'routingResult';
-        
-        trRoutingProcessManager.status = jest.fn(() => Promise.resolve({
-            status: 'started'
-        } as any));
-        TransitAccessibilityMapCalculator.calculate = jest.fn(() => Promise.resolve({
-            routingResult: routingResult
-        } as any));
+        const attributes = { attribute: 'value' };
+        const result = 'result';
 
-        const response = await request(app).post('/api/accessibility?withGeojson=false');
+        (calculateAccessibilityMap as jest.Mock).mockResolvedValue(result);
 
-        expect(response.status).toStrictEqual(200);
-        expect(response.body).toStrictEqual({resultByNode: routingResult});
-        expect(TransitAccessibilityMapCalculator.calculate).toBeCalled();
-    });
-
-    test('POST /api/accessibility, with geojson', async () => {
-        const result = {
-            polygons: 'polygons',
-            strokes: 'strokes',
-            resultByNode: 'resultByNode',
-        };
-        
-        trRoutingProcessManager.status = jest.fn(() => Promise.resolve({
-            status: 'started'
-        } as any));
-        transitObjectDataHandlers.nodes.geojsonCollection! = jest.fn(() => Promise.resolve({
-            geojson: {
-                features: 'features'
-            }
-        }));
-        TransitAccessibilityMapCalculator.calculateWithPolygons = jest.fn(() => Promise.resolve(result as any));
-
-        const response = await request(app).post('/api/accessibility?withGeojson=true');
+        const response = await request(app)
+            .post('/api/accessibility?withGeojson=false')
+            .send(attributes);
 
         expect(response.status).toStrictEqual(200);
         expect(response.body).toStrictEqual(result);
-        expect(transitObjectDataHandlers.nodes.geojsonCollection).toBeCalled();
-        expect(TransitAccessibilityMapCalculator.calculateWithPolygons).toBeCalled();
+        expect(calculateAccessibilityMap).toBeCalledWith(attributes, false);
+    });
+
+    test('POST /api/accessibility, with geojson', async () => {
+        const attributes = { attribute: 'value' };
+        const result = 'result';
+
+        (calculateAccessibilityMap as jest.Mock).mockResolvedValue(result);
+
+        const response = await request(app)
+            .post('/api/accessibility?withGeojson=true')
+            .send(attributes);
+
+        expect(response.status).toStrictEqual(200);
+        expect(response.body).toStrictEqual(result);
+        expect(calculateAccessibilityMap).toBeCalledWith(attributes, true);
     });
 
     test('POST /api/accessibility, with error', async () => {
-        trRoutingProcessManager.status = jest.fn(() => Promise.resolve({
-            status: 'started'
-        } as any));
-        TransitAccessibilityMapCalculator.calculate = jest.fn(() => {
+        (calculateAccessibilityMap as jest.Mock).mockImplementation(() => {
             throw new Error();
         });
 
-        const response = await request(app).post('/api/accessibility?withGeojson=false');
+        const response = await request(app).post('/api/accessibility');
 
         expect(response.status).toStrictEqual(500);
-        expect(TransitAccessibilityMapCalculator.calculate).toBeCalled();
+        expect(calculateAccessibilityMap).toBeCalled();
     });
 
     test('POST /api/route, without geojson', async () => {
-        const transitResult = 'transitResult';
-        const walkingResult = 'walkingResult';
-        const drivingResult = 'drivingResult';
+        const attributes = { attribute: 'value' };
+        const result = 'result';
 
-        const resultsByMode = {
-            transit: {
-                getParams: () => transitResult
-            },
-            walking: {
-                getParams: () => walkingResult
-            },
-            driving: {
-                getParams: () => drivingResult
-            },
-        }
-        
-        trRoutingProcessManager.status = jest.fn().mockResolvedValue({
-            status: 'started'
-        } as any);
-        TransitRoutingCalculator.calculate = jest.fn().mockResolvedValue(resultsByMode);
+        (calculateRoute as jest.Mock).mockResolvedValue(result);
 
-        const response = await request(app).post('/api/route?withGeojson=false');
+        const response = await request(app)
+            .post('/api/route?withGeojson=false')
+            .send(attributes);
 
         expect(response.status).toStrictEqual(200);
-        expect(response.body).toStrictEqual({
-            transit: transitResult,
-            walking: walkingResult,
-            driving: drivingResult
-        });
-        expect(TransitRoutingCalculator.calculate).toBeCalled();
+        expect(response.body).toStrictEqual(result);
+        expect(calculateRoute).toBeCalledWith(attributes, false);
     });
 
     test('POST /api/route, with geojson', async () => {
-        const transitResult = 'transitResult';
-        const walkingResult = 'walkingResult';
-        const drivingResult = 'drivingResult';
-        const transitGeojson = 'transitGeojson';
-        const walkingGeojson = 'walkingGeojson';
-        const drivingGeojson = 'drivingGeojson';
+        const attributes = { attribute: 'value' };
+        const result = 'result';
 
-        const resultsByMode = {
-            transit: {
-                getParams: () =>  {
-                    return { result: transitResult }
-                },
-                getAlternativesCount: () => 1,
-                getPathGeojson: async () => transitGeojson
-            },
-            walking: {
-                getParams: () => {
-                    return { result: walkingResult }
-                },
-                getAlternativesCount: () => 1,
-                getPathGeojson: async () => walkingGeojson
-            },
-            driving: {
-                getParams: () => {
-                    return { result: drivingResult }
-                },
-                getAlternativesCount: () => 1,
-                getPathGeojson: async () => drivingGeojson
-            },
-        }
-        
-        trRoutingProcessManager.status = jest.fn().mockResolvedValue({
-            status: 'started'
-        } as any);
-        TransitRoutingCalculator.calculate = jest.fn().mockResolvedValue(resultsByMode);
-        transitObjectDataHandlers.paths.geojsonCollection! = jest.fn().mockResolvedValue({
-            geojson: {
-                features: 'features'
-            }
-        });
+        (calculateRoute as jest.Mock).mockResolvedValue(result);
 
-        const response = await request(app).post('/api/route?withGeojson=true');
+        const response = await request(app)
+            .post('/api/route?withGeojson=true')
+            .send(attributes);
 
         expect(response.status).toStrictEqual(200);
-        expect(response.body).toStrictEqual({
-            transit: {
-                result: transitResult,
-                pathsGeojson: [transitGeojson]
-            },
-            walking: {
-                result: walkingResult,
-                pathsGeojson: [walkingGeojson]
-            },
-            driving: {
-                result: drivingResult,
-                pathsGeojson: [drivingGeojson]
-            }
-        });
-        expect(TransitRoutingCalculator.calculate).toBeCalled();
+        expect(response.body).toStrictEqual(result);
+        expect(calculateRoute).toBeCalledWith(attributes, true);
     });
 
-    test('POST /api/route, with geojson and alternatives', async () => {
-        const transitResult = 'transitResult';
-        const walkingResult = 'walkingResult';
-        const drivingResult = 'drivingResult';
-        const transitGeojson = 'transitGeojson';
-        const walkingGeojson = 'walkingGeojson';
-        const drivingGeojson = 'drivingGeojson';
-
-        const resultsByMode = {
-            transit: {
-                getParams: () =>  {
-                    return { result: transitResult }
-                },
-                getAlternativesCount: () => 3,
-                getPathGeojson: async () => transitGeojson
-            },
-            walking: {
-                getParams: () => {
-                    return { result: walkingResult }
-                },
-                getAlternativesCount: () => 2,
-                getPathGeojson: async () => walkingGeojson
-            },
-            driving: {
-                getParams: () => {
-                    return { result: drivingResult }
-                },
-                getAlternativesCount: () => 4,
-                getPathGeojson: async () => drivingGeojson
-            },
-        }
-        
-        trRoutingProcessManager.status = jest.fn().mockResolvedValue({
-            status: 'started'
-        } as any);
-        TransitRoutingCalculator.calculate = jest.fn().mockResolvedValue(resultsByMode);
-        transitObjectDataHandlers.paths.geojsonCollection! = jest.fn().mockResolvedValue({
-            geojson: {
-                features: 'features'
-            }
-        });
-
-        const response = await request(app).post('/api/route?withGeojson=true');
-
-        expect(response.status).toStrictEqual(200);
-        expect(response.body).toStrictEqual({
-            transit: {
-                result: transitResult,
-                pathsGeojson: [transitGeojson, transitGeojson, transitGeojson]
-            },
-            walking: {
-                result: walkingResult,
-                pathsGeojson: [walkingGeojson, walkingGeojson]
-            },
-            driving: {
-                result: drivingResult,
-                pathsGeojson: [drivingGeojson, drivingGeojson, drivingGeojson, drivingGeojson]
-            }
-        });
-        expect(TransitRoutingCalculator.calculate).toBeCalled();
-    });
-
-    test('POST /api/route, with error', async () => {       
-        trRoutingProcessManager.status = jest.fn().mockResolvedValue({
-            status: 'started'
-        } as any);
-        TransitRoutingCalculator.calculate = jest.fn(() => {
+    test('POST /api/route, with error', async () => {
+        (calculateRoute as jest.Mock).mockImplementation(() => {
             throw new Error();
         });
 
         const response = await request(app).post('/api/route');
 
         expect(response.status).toStrictEqual(500);
-        expect(TransitRoutingCalculator.calculate).toBeCalled();
+        expect(calculateRoute).toBeCalled();
     });
 });
